+++ Mods ChoGGi/Mark Deposit Ground/metadata.lua
@@ -7,7 +7,8 @@ return PlaceObj("ModDef", {
 	"image", "Preview.png",
 	"id", "ChoGGi_MarkDepositGround",
 	"steam_id", "1555446081",
-	"pops_any_uuid", "8a57f30a-6bae-46c0-99b9-ca497bf74bd3",
+--~ 	"pops_desktop_uuid", "8a57f30a-6bae-46c0-99b9-ca497bf74bd3",
+	"pops_any_uuid", "3b37cc32-66d0-4233-8366-9038691ebe0e",
 	"author", "ChoGGi",
 	"lua_revision", 245618,
 	"code", {